feat: Changes to Open Controller configuration format...
Open Controller configuration format is now unified for all of its
services. This means that to run Clockwork, Traffic Indicators and
Simengine, you need only one file.

Note that this change only addresses Clockwork, and other parts need to
be updated later.
